+++ derive/src/simple_object.rs
@@ -3,16 +3,19 @@ use proc_macro::TokenStream;
 use quote::quote;
 use std::str::FromStr;
 use syn::ext::IdentExt;
-use syn::{Error, Ident, Type};
+use syn::{Error, Ident, Path, Type};
 
 use crate::args::{self, RenameRuleExt, RenameTarget, SimpleObjectField};
 use crate::utils::{
     gen_deprecation, generate_guards, get_crate_name, get_rustdoc, visible_fn, GeneratorResult,
 };
 
+#[derive(Debug)]
 struct DerivedFieldMetadata {
     ident: Ident,
     into: Type,
+    owned: Option<bool>,
+    with: Option<Path>,
 }
 
 struct SimpleObjectFieldGenerator<'a> {
@@ -60,10 +63,25 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
         for derived in &field.derived {
             if derived.name.is_some() && derived.into.is_some() {
                 let name = derived.name.clone().unwrap();
-                let into = Type::Verbatim(
+                let into = match syn::parse2::<Type>(
                     proc_macro2::TokenStream::from_str(&derived.into.clone().unwrap()).unwrap(),
-                );
-                let derived = DerivedFieldMetadata { ident: name, into };
+                ) {
+                    Ok(e) => e,
+                    _ => {
+                        return Err(Error::new_spanned(
+                            &name,
+                            "derived into must be a valid type.",
+                        )
+                        .into());
+                    }
+                };
+
+                let derived = DerivedFieldMetadata {
+                    ident: name,
+                    into,
+                    owned: derived.owned,
+                    with: derived.with.clone(),
+                };
 
                 processed_fields.push(SimpleObjectFieldGenerator {
                     field,
@@ -83,7 +101,6 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
             None => return Err(Error::new_spanned(&ident, "All fields must be named.").into()),
         };
 
-        let is_derived = derived.is_some();
         let ident = if let Some(derived) = derived {
             &derived.ident
         } else {
@@ -109,10 +126,17 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
             None => quote! { ::std::option::Option::None },
         };
         let vis = &field.vis;
+
         let ty = if let Some(derived) = derived {
             &derived.into
         } else {
             &field.ty
+        };
+
+        let owned = if let Some(derived) = derived {
+            derived.owned.unwrap_or(field.owned)
+        } else {
+            field.owned
         };
 
         let cache_control = {
@@ -152,7 +176,9 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
             |guard| quote! { #guard.check(ctx).await.map_err(|err| err.into_server_error(ctx.item.pos))?; },
         );
 
-        let mut block = match !field.owned {
+        let with_function = derived.as_ref().and_then(|x| x.with.as_ref());
+
+        let mut block = match !owned {
             true => quote! {
                 &self.#base_ident
             },
@@ -161,16 +187,17 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
             },
         };
 
-        block = match is_derived {
-            false => quote! {
-                #block
+        block = match (derived, with_function) {
+            (Some(_), Some(with)) => quote! {
+                #with(#block)
             },
-            true => quote! {
+            (Some(_), None) => quote! {
                 ::std::convert::Into::into(#block)
             },
+            (_, _) => block,
         };
 
-        let ty = match !field.owned {
+        let ty = match !owned {
             true => quote! { &#ty },
             false => quote! { #ty },
         };
@@ -266,6 +293,7 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
                         extends: #extends,
                         keys: ::std::option::Option::None,
                         visible: #visible,
+                        is_subscription: false,
                     })
                 }
             }
@@ -316,6 +344,7 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
                         extends: #extends,
                         keys: ::std::option::Option::None,
                         visible: #visible,
+                        is_subscription: false,
                     })
                 }