+++ integrations/axum/src/response.rs
@@ -1,5 +1,3 @@
-use std::convert::TryFrom;
-
 use axum::body::Body;
 use axum::response::IntoResponse;
 use headers::HeaderName;