+++ parser/src/parse/mod.rs
@@ -312,3 +312,21 @@ fn parse_name(pair: Pair<Rule>, pc: &mut PositionCalculator) -> Result<Positione
     debug_assert_eq!(pair.as_rule(), Rule::name);
     Ok(Positioned::new(Name::new(pair.as_str()), pc.step(&pair)))
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_number_lookahead_restrictions() {
+        GraphQLParser::parse(Rule::const_list, "[123 abc]").unwrap();
+        GraphQLParser::parse(Rule::const_list, "[123.0123 abc]").unwrap();
+        GraphQLParser::parse(Rule::const_list, "[123.0123e7 abc]").unwrap();
+        GraphQLParser::parse(Rule::const_list, "[123.0123e77 abc]").unwrap();
+
+        assert!(GraphQLParser::parse(Rule::const_list, "[123abc]").is_err());
+        assert!(GraphQLParser::parse(Rule::const_list, "[123.0123abc]").is_err());
+        assert!(GraphQLParser::parse(Rule::const_list, "[123.0123e7abc]").is_err());
+        assert!(GraphQLParser::parse(Rule::const_list, "[123.0123e77abc]").is_err());
+    }
+}