+++ src/types/id.rs
@@ -1,4 +1,3 @@
-use std::convert::TryFrom;
 use std::num::ParseIntError;
 use std::ops::{Deref, DerefMut};