+++ value/src/lib.rs
@@ -10,9 +10,7 @@ mod value_serde;
 mod variables;
 
 use std::borrow::{Borrow, Cow};
-use std::convert::{TryFrom, TryInto};
 use std::fmt::{self, Display, Formatter, Write};
-use std::iter::FromIterator;
 use std::ops::Deref;
 use std::sync::Arc;